+++ asset/css/write.css
@@ -1,3 +1,9 @@
+.write-container {
+  width: 100%;
+  height: 100%;
+  background-color: var(--white);
+}
+
 .write-content-container {
   padding: 24px 16px 16px 0;
 }
